requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
nibabel==3.0.2
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pkg-resources==0.0.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-probability==0.8.0
termcolor==1.1.0
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
